olympus/model_engine/requirements.txt
torch
accelerate
transformers
diffusers
vllm
click
pydantic
numpy
tqdm
tensorboard
wandb
prometheus-client>=0.17.1
pyyaml>=6.0.1
